DS-261: Expandable unit test with error summary
